+++ apps/workflows-dashboard/src/domains/workflows/api/workflow-metrics/workflow-metrics.types.ts
@@ -17,5 +17,5 @@ export type ICasesPerStatusStats = Record<IWorkflowStatus, number>;
 
 export interface GetCasesPerStatusDto {
   // UNIX timestamp
-  fromDate: number;
+  fromDate?: number;
 }