+++ packages/common/src/schemas/documents/workflow/documents/schemas/GH.ts
@@ -461,6 +461,42 @@ export const ghanaDocuments: TDocument[] = [
     },
   },
   {
+    category: 'proof_of_employment',
+    type: 'form_a',
+    issuer: {
+      type: 'private',
+      country: 'GH',
+    },
+    issuingVersion: 1,
+    version: 1,
+    propertiesSchema: {
+      type: 'object',
+      properties: {
+        nationalIdNumber: {
+          type: 'string',
+          pattern: ghNationalIdNumber,
+        },
+        docNumber: {
+          type: 'string',
+          pattern: alphaNumeric,
+        },
+        employeeName: {
+          type: 'string',
+        },
+        position: {
+          type: 'string',
+        },
+        employerName: {
+          type: 'string',
+        },
+        issuingDate: {
+          type: 'string',
+          format: 'date',
+        },
+      },
+    },
+  },
+  {
     category: 'proof_of_address',
     type: 'form_a',
     issuer: {
@@ -494,6 +530,42 @@ export const ghanaDocuments: TDocument[] = [
     },
   },
   {
+    category: 'proof_of_employment',
+    type: 'form_3',
+    issuer: {
+      type: 'private',
+      country: 'GH',
+    },
+    issuingVersion: 1,
+    version: 1,
+    propertiesSchema: {
+      type: 'object',
+      properties: {
+        nationalIdNumber: {
+          type: 'string',
+          pattern: ghNationalIdNumber,
+        },
+        docNumber: {
+          type: 'string',
+          pattern: alphaNumeric,
+        },
+        employeeName: {
+          type: 'string',
+        },
+        position: {
+          type: 'string',
+        },
+        employerName: {
+          type: 'string',
+        },
+        issuingDate: {
+          type: 'string',
+          format: 'date',
+        },
+      },
+    },
+  },
+  {
     category: 'proof_of_address',
     type: 'form_3',
     issuer: {
@@ -517,6 +589,75 @@ export const ghanaDocuments: TDocument[] = [
           type: 'string',
         },
         physicalAddress: {
+          type: 'string',
+        },
+        issuingDate: {
+          type: 'string',
+          format: 'date',
+        },
+      },
+    },
+  },
+  {
+    category: 'proof_of_address',
+    type: 'form_4',
+    issuer: {
+      type: 'government',
+      country: 'GH',
+    },
+    issuingVersion: 1,
+    version: 1,
+    propertiesSchema: {
+      type: 'object',
+      properties: {
+        nationalIdNumber: {
+          type: 'string',
+          pattern: ghNationalIdNumber,
+        },
+        docNumber: {
+          type: 'string',
+          pattern: alphaNumeric,
+        },
+        userAddress: {
+          type: 'string',
+        },
+        physicalAddress: {
+          type: 'string',
+        },
+        issuingDate: {
+          type: 'string',
+          format: 'date',
+        },
+      },
+    },
+  },
+  {
+    category: 'proof_of_employment',
+    type: 'form_4',
+    issuer: {
+      type: 'private',
+      country: 'GH',
+    },
+    issuingVersion: 1,
+    version: 1,
+    propertiesSchema: {
+      type: 'object',
+      properties: {
+        nationalIdNumber: {
+          type: 'string',
+          pattern: ghNationalIdNumber,
+        },
+        docNumber: {
+          type: 'string',
+          pattern: alphaNumeric,
+        },
+        employeeName: {
+          type: 'string',
+        },
+        position: {
+          type: 'string',
+        },
+        employerName: {
           type: 'string',
         },
         issuingDate: {