+++ services/workflows-service/src/metrics/repository/metrics.repository.ts
@@ -19,7 +19,7 @@ import { IAggregateAverageResolutionTime } from '@/metrics/repository/types/aggr
 import { GetUserAverageAssignmentTimeParams } from '@/metrics/repository/types/get-user-average-assignment-time.params';
 import { IAggregateAverageAssignmentTime } from '@/metrics/repository/types/aggregate-average-assignment-time';
 import { GetUserAverageReviewTimeParams } from '@/metrics/repository/types/get-user-average-review-time.params';
-import { UserAverageReviewTimeModel } from '@/metrics/repository/models/user-average-review-time.model';
+import { AverageReviewTimeModel } from '@/metrics/repository/models/average-review-time.model';
 import { IAggregateAverageReviewTime } from '@/metrics/repository/types/aggregate-average-review-time';
 import { ListUserCasesResolvedDailyParams } from '@/metrics/repository/types/list-user-cases-resolved-daily.params';
 import { CasesResolvedInDay } from '@/metrics/repository/models/cases-resolved-daily.model';
@@ -107,10 +107,11 @@ export class MetricsRepository {
     const results = await this.prismaService.$queryRawUnsafe<IAggregateApprovalRate[]>(
       aggregateApprovalRateQuery,
       params.fromDate,
-      params.userId,
     );
 
-    return results.length ? plainToClass(ApprovalRateModel, results.at(-1)) : null;
+    return results.length
+      ? plainToClass(ApprovalRateModel, { approvalRate: results.at(-1)?.approvalRate })
+      : null;
   }
 
   async getUserAverageResolutionTime(
@@ -119,10 +120,11 @@ export class MetricsRepository {
     const results = await this.prismaService.$queryRawUnsafe<IAggregateAverageResolutionTime[]>(
       aggregateAverageResolutionTimeQuery,
       params.fromDate,
-      params.userId,
     );
 
-    return results.length ? plainToClass(AverageResolutionTimeModel, results.at(-1)) : null;
+    return results.length
+      ? plainToClass(AverageResolutionTimeModel, { time: results.at(-1)?.average_time })
+      : null;
   }
 
   async getUserAverageAssignmentTime(
@@ -131,22 +133,24 @@ export class MetricsRepository {
     const results = await this.prismaService.$queryRawUnsafe<IAggregateAverageAssignmentTime[]>(
       aggregateAverageAssignmentTimeQuery,
       params.fromDate,
-      params.userId,
     );
 
-    return results.length ? plainToClass(AverageAssignmentTimeModel, results.at(-1)) : null;
+    return results.length
+      ? plainToClass(AverageAssignmentTimeModel, { time: results.at(-1)?.average_time })
+      : null;
   }
 
   async getUserAverageReviewTime(
     params: GetUserAverageReviewTimeParams,
-  ): Promise<UserAverageReviewTimeModel | null> {
+  ): Promise<AverageReviewTimeModel | null> {
     const results = await this.prismaService.$queryRawUnsafe<IAggregateAverageReviewTime[]>(
       aggregateAverageReviewTimeQuery,
       params.fromDate,
-      params.userId,
     );
 
-    return results.length ? plainToClass(UserAverageReviewTimeModel, results.at(-1)) : null;
+    return results.length
+      ? plainToClass(AverageReviewTimeModel, { time: results.at(-1)?.average_time })
+      : null;
   }
 
   async listCasesResolvedDaily(