+++ services/workflows-service/src/metrics/repository/models/average-resolution-time.model.ts
@@ -3,6 +3,6 @@ import { Transform } from 'class-transformer';
 
 export class AverageResolutionTimeModel {
   @ApiProperty({ description: 'Average resolution time in milliseconds.' })
-  @Transform(({ value }) => (value === null ? 0 : value))
-  time!: number;
+  @Transform(({ value }) => (!value ? 0 : value.split('.')[0]))
+  time!: string;
 }