+++ services/workflows-service/src/metrics/repository/sql/aggregate-average-assignment-time.sql.ts
@@ -1,18 +1,20 @@
 export const aggregateAverageAssignmentTimeQuery = `
-select
-	avg(
-	(select sum(case
-		when "createdAt" notnull and "assignedAt" notnull
-		then (extract(epoch from "assignedAt") - extract(epoch from "createdAt")) * 1000
-		else 0
-		end)::int as time
-	from "WorkflowRuntimeData"
-	where "assigneeId" = "User".id
-	and "assignedAt" >= $1)
-	)::int as time
-from
-	"User"
-where
-	"User".id = coalesce($2,
-	"id")
+SELECT AVG(time)::varchar as average_time
+FROM (
+  SELECT
+    "User".id,
+    COALESCE(
+      EXTRACT(EPOCH FROM ("WorkflowRuntimeData"."assignedAt" - "WorkflowRuntimeData"."createdAt")) * 1000,
+      0
+    ) AS time
+  FROM
+    "User"
+  LEFT JOIN
+    "WorkflowRuntimeData"
+    ON "User".id = "WorkflowRuntimeData"."assigneeId"
+  WHERE
+    "WorkflowRuntimeData"."createdAt" IS NOT NULL
+    AND
+    "WorkflowRuntimeData"."assignedAt" >= $1
+) AS T
 `;