+++ services/workflows-service/src/metrics/service/models/user-workflow-processing-statistic.model.ts
@@ -5,14 +5,14 @@ export class UserWorkflowProcessingStatisticModel {
   id!: string;
 
   @ApiProperty()
-  approvalRate!: number;
+  approvalRate!: string;
 
   @ApiProperty({ description: 'Average resolution time in milliseconds' })
-  averageResolutionTime!: number;
+  averageResolutionTime!: string;
 
   @ApiProperty({ description: 'Average assignment time in milliseconds' })
-  averageAssignmentTime!: number;
+  averageAssignmentTime!: string;
 
   @ApiProperty({ description: 'Average review time in milliseconds' })
-  averageReviewTime!: number;
+  averageReviewTime!: string;
 }