+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/workflow-core
 
+## 0.4.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.6
+
+## 0.4.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.5
+
 ## 0.4.14
 
 ### Patch Changes